Added code to address dependencies and chain-deletion of cache entries.
Rev: lib/modules/Cache.pmod/Policy.pmod/Timed.pike:1.4
Rev: lib/modules/Cache.pmod/Storage.pmod/Base.pike:1.5
Rev: lib/modules/Cache.pmod/Storage.pmod/Gdbm.pike:1.6
Rev: lib/modules/Cache.pmod/Storage.pmod/Memory.pike:1.5
Rev: lib/modules/Cache.pmod/Storage.pmod/Yabu.pike:1.7
Rev: lib/modules/Cache.pmod/cache.pike:1.5

+++ lib/modules/Cache.pmod/Policy.pmod/Timed.pike
@@ -3,7 +3,7 @@
  * by Francesco Chemolli <kinkie@roxen.com>
  * (C) 2000 Roxen IS
  *
- * $Id: Timed.pike,v 1.3 2000/09/28 03:38:30 hubbe Exp $
+ * $Id: Timed.pike,v 1.4 2001/01/01 22:49:35 kinkie Exp $
  */
 
 #pike __REAL_VERSION__
@@ -23,11 +23,17 @@ void expire(Cache.Storage storage) {
   string key=storage->first();
   while (key) {
     Cache.Data got=storage->get(key,1);
-    werror(sprintf("examining '%s' (age: %d, e: %d, v: %f)\n",
-                   key, now-got->atime, got->etime, got->cost));
+    if (!objectp(got)) {
+      //if got==0, there must have been some
+      //dependents acting. This is kludgy... :-(
+      key=storage->next();
+      continue;
+    }
     if (got->atime < limit ||
         (got->etime && got->etime < now) ) {
-      werror("deleting\n");
+      werror("deleting %s (age: %d, now: %d, etime: %d)\n",
+             key, now - got->atime, 
+             now, got->etime);
       storage->delete(key);
     }
     key=storage->next();
